requirements.txt
fastapi
uvicorn
beautifulsoup4
mysql-connector-python
pydantic
requests
selenium
